requirements.txt
appdirs==1.4.3
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
cycler==0.10.0
decorator==4.0.11
dill==0.2.6
enum34==1.1.6
h5py==2.7.0
html5lib==0.9999999
idna==2.6
ipython==6.0.0
ipython-genutils==0.2.0
jedi==0.10.2
jsonschema==2.6.0
jupyter-core==4.4.0
Keras==2.1.1
Markdown==2.6.9
matplotlib==2.0.2
nbformat==4.4.0
nibabel==2.2.1
numpy==1.13.3
olefile==0.44
packaging==16.8
pandas==0.20.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
plotly==2.5.0
prompt-toolkit==1.0.14
protobuf==3.5.0.post1
ptyprocess==0.5.1
Pygments==2.2.0
pynrrd==0.2.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
requests==2.18.4
scikit-learn==0.18.1
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorflow==1.4.0
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0rc3
Theano==0.9.0
traitlets==4.3.2
typing==3.6.1
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2

